feat: expand morphology operations for annotations
Now, all morphology operations are supported for annotations. A new 'labelmask_multiupdate()' function was created to store simutaneously the writing or erasing of different dots in the image as one operation only. This functionality is needed by some morphology operations as closing, opening and smooth. This way the 'undo' button funcitons correctly.