feat: add fill holes + different structuring elements
The structuring elements and operations are set for 2D operations only, because they operate on annotations that are always made in 2D slices. The fill holes operation needs the input to be reduced for int8 because this function can't be broken to execute in chunks. This allows the operation to be performed in only one GPU even for the biggest expected input images.